Move OneDriveUrlProvider to GraphUrlProvider, sharing implementation for all Graph protocols
