+++ onedrive/src/main/java/ch/cyberduck/core/onedrive/GraphSession.java
@@ -15,12 +15,7 @@ package ch.cyberduck.core.onedrive;
  * GNU General Public License for more details.
  */
 
-import ch.cyberduck.core.Credentials;
-import ch.cyberduck.core.DefaultIOExceptionMappingService;
-import ch.cyberduck.core.Host;
-import ch.cyberduck.core.HostKeyCallback;
-import ch.cyberduck.core.LoginCallback;
-import ch.cyberduck.core.Path;
+import ch.cyberduck.core.*;
 import ch.cyberduck.core.exception.BackgroundException;
 import ch.cyberduck.core.exception.HostParserException;
 import ch.cyberduck.core.features.AttributesFinder;
@@ -41,19 +36,7 @@ import ch.cyberduck.core.http.HttpSession;
 import ch.cyberduck.core.oauth.OAuth2AuthorizationService;
 import ch.cyberduck.core.oauth.OAuth2ErrorResponseInterceptor;
 import ch.cyberduck.core.oauth.OAuth2RequestInterceptor;
-import ch.cyberduck.core.onedrive.features.GraphAttributesFinderFeature;
-import ch.cyberduck.core.onedrive.features.GraphCopyFeature;
-import ch.cyberduck.core.onedrive.features.GraphDeleteFeature;
-import ch.cyberduck.core.onedrive.features.GraphDirectoryFeature;
-import ch.cyberduck.core.onedrive.features.GraphFileIdProvider;
-import ch.cyberduck.core.onedrive.features.GraphFindFeature;
-import ch.cyberduck.core.onedrive.features.GraphMoveFeature;
-import ch.cyberduck.core.onedrive.features.GraphPromptUrlProvider;
-import ch.cyberduck.core.onedrive.features.GraphQuotaFeature;
-import ch.cyberduck.core.onedrive.features.GraphReadFeature;
-import ch.cyberduck.core.onedrive.features.GraphTimestampFeature;
-import ch.cyberduck.core.onedrive.features.GraphTouchFeature;
-import ch.cyberduck.core.onedrive.features.GraphWriteFeature;
+import ch.cyberduck.core.onedrive.features.*;
 import ch.cyberduck.core.proxy.Proxy;
 import ch.cyberduck.core.proxy.ProxyFactory;
 import ch.cyberduck.core.shared.BufferWriteFeature;
@@ -252,6 +235,9 @@ public abstract class GraphSession extends HttpSession<OneDriveAPI> {
         }
         if(type == Quota.class) {
             return (T) new GraphQuotaFeature(this, fileid);
+        }
+        if(type == UrlProvider.class) {
+            return (T) new GraphUrlProvider();
         }
         if(type == PromptUrlProvider.class) {
             return (T) new GraphPromptUrlProvider(this);
